backend/test file/test_full_auth.py
#!/usr/bin/env python3
"""
Comprehensive test for authentication flow
"""

import requests
import json
from urllib3.exceptions import InsecureRequestWarning
import urllib3

# Disable SSL warnings for self-signed certificates
urllib3.disable_warnings(InsecureRequestWarning)

def test_auth_flow():
    print("🔍 Testing Authentication Flow...")
    print()
    
    # Test 1: Root endpoint
    try:
        response = requests.get("https://localhost:8443/", verify=False, timeout=10)
        if response.status_code == 200:
            print("✅ Root endpoint working")
            data = response.json()
            if "/auth/login" in str(data.get("endpoints", {})):
                print("✅ Auth endpoint is listed in available endpoints")
            else:
                print("❌ Auth endpoint not found in available endpoints")
        else:
            print(f"❌ Root endpoint failed: {response.status_code}")
    except Exception as e:
        print(f"❌ Root endpoint error: {e}")
        return False
    
    print()
    
    # Test 2: CORS headers
    try:
        response = requests.options("https://localhost:8443/auth/login", 
                                   verify=False, 
                                   timeout=10,
                                   headers={
                                       "Origin": "http://localhost:5173",
                                       "Access-Control-Request-Method": "POST",
                                       "Access-Control-Request-Headers": "Content-Type"
                                   })
        print(f"CORS preflight status: {response.status_code}")
        cors_headers = {k: v for k, v in response.headers.items() if 'access-control' in k.lower()}
        if cors_headers:
            print("✅ CORS headers present:")
            for header, value in cors_headers.items():
                print(f"   {header}: {value}")
        else:
            print("❌ No CORS headers found")
    except Exception as e:
        print(f"⚠️ CORS test error: {e}")
    
    print()
    
    # Test 3: Auth endpoint structure
    try:
        response = requests.post("https://localhost:8443/auth/login", 
                               json={"id_token": "invalid_token"}, 
                               headers={"Content-Type": "application/json"},
                               verify=False, 
                               timeout=10)
        
        print(f"Auth endpoint status: {response.status_code}")
        print(f"Response headers: {dict(response.headers)}")
        print(f"Response body: {response.text}")
        
        if response.status_code in [401, 422]:
            print("✅ Auth endpoint is properly rejecting invalid tokens")
        else:
            print(f"❌ Unexpected status code: {response.status_code}")
            
    except Exception as e:
        print(f"❌ Auth endpoint test error: {e}")
        return False
    
    print()
    
    # Test 4: Check if Google OAuth verification is working
    print("🔍 Testing Google OAuth integration...")
    try:
        from google.oauth2 import id_token
        from google.auth.transport import requests as google_requests
        print("✅ Google OAuth libraries are available")
        
        # Test with the actual Google Client ID
        client_id = "761585986838-odetbtp33g2gtrp82ikrvmbt51r5f250.apps.googleusercontent.com"
        print(f"✅ Using Google Client ID: {client_id[:20]}...")
        
    except ImportError as e:
        print(f"❌ Google OAuth libraries missing: {e}")
        return False
    except Exception as e:
        print(f"⚠️ Google OAuth test error: {e}")
    
    print()
    print("🔍 Summary:")
    print("- Backend is running on HTTPS ✅")
    print("- Auth endpoint is accessible ✅") 
    print("- CORS should be configured ✅")
    print("- Google OAuth libraries are available ✅")
    print()
    print("💡 If frontend login is still failing, check:")
    print("1. Browser console for specific error messages")
    print("2. Network tab to see the actual request/response")
    print("3. Accept the self-signed certificate in browser")
    print("4. Make sure frontend is sending valid Google ID token")
    
    return True

if __name__ == "__main__":
    test_auth_flow()
